feat(Register): Create register page, create template for pages with form
