+++ App.tsx
@@ -1,8 +1,9 @@
 import { NativeRouter, Routes, Route } from "react-router-native";
 
 import GlobalProvider from "~/Contexts/globalContext";
+import Login from "~/Views/Account/Login";
+import Register from "~/Views/Account/Register";
 import Home from "~/Views/Home";
-import Login from "~/Views/Login";
 import Page from "~/Views/Page";
 
 export default function App() {
@@ -13,6 +14,7 @@ export default function App() {
           <Routes>
             <Route index element={<Home />} />
             <Route path="/login" element={<Login />} />
+            <Route path="/register" element={<Register />} />
           </Routes>
         </Page>
       </GlobalProvider>
