+++ src/Components/Form/Input/styles.ts
@@ -17,7 +17,7 @@ export const InputElement = styled.TextInput`
   border-color: ${({ error }: InputStyleProps) =>
     error ? color?.semantic?.dangerText : color?.neutral?.black20};
   border-radius: ${shape?.radius?.input}px;
-  padding: 13px 16px;
+  padding: 8px 16px;
   background-color: ${color?.primary?.white};
   margin: 10px 0 5px;
 `;
@@ -28,7 +28,7 @@ export const Select = styled.View`
   border-color: ${color?.neutral?.black20};
   overflow: hidden;
   border-radius: ${shape?.radius?.input}px;
-  padding: 3px 16px;
+  padding: 0 16px;
   background-color: ${color?.primary?.white};
   margin: 10px 0 5px;
 `;
