+++ src/Components/Form/index.tsx
@@ -29,7 +29,6 @@ export default function Form({
   const [ctaDisabled, setCtaDisabled] = useState(true);
 
   useEffect(() => {
-    alert(JSON.stringify(formData));
     const dataValues = Object.values(formData);
     if (dataValues?.length === 0 || dataValues?.includes(null)) {
       setCtaDisabled(true);
