+++ src/Views/Account/Login/styles.ts
@@ -1,0 +1,5 @@
+import styled from "styled-components/native";
+
+export const BottomText = styled.View`
+  margin: 16px 0 10px;
+`;
